backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
appnope==0.1.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
clikit==0.6.2
cmdstanpy==0.9.68
convertdate==2.4.0
crashtest==0.3.1
cycler==0.11.0
Cython==0.29.30
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
ephem==4.1.3
Flask==2.1.2
fonttools==4.33.3
frozenlist==1.3.0
hijri-converter==2.2.3
holidays==0.13
httpstan==4.6.1
idna==3.3
importlib-metadata==4.11.3
influxdb-client==1.28.0
ipykernel==6.13.0
ipython==7.32.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.0
jupyter-core==4.10.0
kiwisolver==1.4.2
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
MarkupSafe==2.1.1
marshmallow==3.15.0
matplotlib==3.5.2
matplotlib-inline==0.1.3
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.21.6
packaging==21.3
paho-mqtt==1.6.1
pandas==1.3.5
parso==0.8.3
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
prophet==1.0.1
psutil==5.9.0
ptyprocess==0.7.0
Pygments==2.12.0
pylev==1.4.0
PyMeeus==0.5.11
pyparsing==3.0.8
pysimdjson==3.2.0
pystan==2.19.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==22.3.0
Rx==3.2.0
schedule==1.1.0
setuptools-git==1.2
six==1.16.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.2.0
ujson==5.3.0
urllib3==1.26.9
wcwidth==0.2.5
webargs==8.1.0
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
